+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 - show jump down button earlier when scrolling up
 - make font-style of info messages consistent with iOS and Android #3034
+- move all verified icons to the end chat/contact name (previously some were still on the avatar image)
 
 ### Fixed
 - fix webxdc: allow `self` and `blob:` in `connect-src` in CSP